+++ docs/changelog.md
@@ -10,6 +10,8 @@ Versioning currently follows `X.Y.Z` where
 
 ## \[Unreleased\]
 
+## \[1.13.15\] - 2025-01-15
+
 ### Added
 
 - Entity `Sample`